update build to support for sqlite3

+++ bootstrap.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env sh
 
-ROOTDIR=`pwd`
+ROOTDIR=$(dirname $(readlink -f $0))
 
 if [ ! -e 'Python-2.7.2.tgz' ]; then
     wget http://www.python.org/ftp/python/2.7.2/Python-2.7.2.tgz
@@ -17,13 +17,14 @@ if [ ! -e 'Python' ]; then
     cd $ROOTDIR/Python
     patch -p1 < ../patch/Python-2.7.2-xcompile.patch
     patch -p1 < ../patch/Python-2.7.2-android.patch
-    cp -r $ROOTDIR/modules/sqlite3/ $ROOTDIR/Python/Modules/
 fi
 
-cd $ROOTDIR/Python-host
-./configure --prefix=$ROOTDIR/prebuilt
-make -j4
-make install
-mv python $ROOTDIR/hostpython
-mv Parser/pgen $ROOTDIR/hostpgen
-make distclean
+if [ ! -e "$ROOTDIR/hostpython" -a ! -e "$ROOTDIR/hostpgen" ]; then
+    cd $ROOTDIR/Python-host
+    ./configure --prefix=$ROOTDIR/prebuilt
+    make -j4
+    make install
+    mv python $ROOTDIR/hostpython
+    mv Parser/pgen $ROOTDIR/hostpgen
+    make distclean
+fi
